Clean up code for homepage buttons
