--- src/components/Hero/HeroInput/heroinput.module.scss
@@ -1,46 +1,0 @@
-.heroInput {
-  outline: none;
-  font-size: 2.4rem;
-  border: none;
-  border-bottom: 2px solid #eaeaea;
-  width: 80%;
-  padding: 5px 0;
-}
-
-.heroInputContainer {
-  position: relative;
-  margin-top: 3.5rem;
-  display: flex;
-}
-
-.heroInputPlaceholder {
-  position: absolute;
-  bottom: 10px;
-  color: var(--seed-text-body);
-  transition: transform 0.2s;
-}
-
-.active {
-  transform: translateY(-35px);
-}
-
-.activeBorder {
-  border-color: var(--seed-color-primary);
-}
-
-.heroButton {
-  padding: 1rem;
-  background-color: var(--seed-color-primary);
-  border-radius: 6px;
-  color: white;
-  border: none;
-  margin-left: 1rem;
-  font-size: 2rem;
-}
-
-.search {
-  display: block;
-  background-color: var(--seed-color-primary);
-  padding: 1rem;
-  cursor: pointer;
-}
