--- src/components/Hero/HeroInput/index.js
@@ -1,1 +1,0 @@
-export { default } from 'components/Hero/HeroInput/HeroInput';
