+++ src/components/Hero/HeroSelect/index.js
@@ -1,0 +1,1 @@
+export { default } from 'components/Hero/HeroSelect/HeroSelect';
